extracted nav, styled a bit
