+++ package.json
@@ -11,17 +11,17 @@
   },
   "devDependencies": {
     "@skeletonlabs/skeleton": "^1.5.0",
-    "@sveltejs/adapter-vercel": "^2.4.3",
-    "@sveltejs/kit": "next",
+    "@sveltejs/adapter-vercel": "^3.0.0",
+    "@sveltejs/kit": "latest",
     "autoprefixer": "^10.4.7",
-    "postcss": "^8.4.14",
+    "postcss": "^8.4.23",
     "postcss-load-config": "^4.0.1",
     "svelte": "^3.55.0",
     "svelte-check": "^3.0.3",
     "svelte-preprocess": "^5.0.3",
-    "tailwindcss": "^3.1.5",
+    "tailwindcss": "^3.3.2",
     "tslib": "^2.3.1",
-    "typescript": "^4.9.4"
+    "typescript": "^5.0.0"
   },
   "type": "module",
   "dependencies": {
