+++ src/routes/+layout.svelte
@@ -2,6 +2,7 @@
   import "@skeletonlabs/skeleton/themes/theme-skeleton.css";
   import "@skeletonlabs/skeleton/styles/skeleton.css";
   import "../app.postcss";
+  import '@skeletonlabs/skeleton/themes/theme-gold-nouveau.css';
 
   import { dev } from "$app/environment";
   import { inject } from "@vercel/analytics";
